# src/main/resources/openapi.yaml
openapi: "3.1.0"
info:
  title: "backend_flashpomo API"
  description: "backend_flashpomo API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
paths:
  /auth/login:
    post:
      summary: "POST auth/login"
      operationId: "login"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
  /auth/register:
    post:
      summary: "POST auth/register"
      operationId: "register"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
  /ensemble/{ensembleId}:
    put:
      summary: "PUT ensemble/{ensembleId}"
      operationId: "updateEnsemble"
      parameters:
        - name: "ensembleId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnsembleRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    delete:
      summary: "DELETE ensemble/{ensembleId}"
      operationId: "deleteEnsemble"
      parameters:
        - name: "ensembleId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET ensemble/{ensembleId}"
      operationId: "getEnsembleById"
      parameters:
        - name: "ensembleId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
  /ensemble/:
    post:
      summary: "POST ensemble/"
      operationId: "createEnsemble"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnsembleRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET ensemble/"
      operationId: "getAllEnsembles"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Ensemble"
...
  /flashcard/ensemble/{ensemble}:
    get:
      summary: "GET flashcard/ensemble/{ensemble}"
      operationId: "getEnsembleById"
      parameters:
        - name: "ensemble"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Flashcard"
...
  /box/{favoriteId}:
    put:
      summary: "PUT box/{favoriteId}"
      operationId: "updateBox"
      parameters:
        - name: "favoriteId"
          in: "path"
          required: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoxRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    delete:
      summary: "DELETE box/{favoriteId}"
      operationId: "deleteBox"
      parameters:
        - name: "favoriteId"
          in: "path"
          required: false
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
  /box/:
    post:
      summary: "POST box/"
      operationId: "createBox"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoxRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET box/"
      operationId: "getAllBoxes"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
  /box/{boxId}:
    get:
      summary: "GET box/{boxId}"
      operationId: "getBoxById"
      parameters:
        - name: "boxId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
  /favorite/{favoriteId}:
    put:
      summary: "PUT favorite/{favoriteId}"
      operationId: "updateFavorite"
      parameters:
        - name: "favoriteId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    delete:
      summary: "DELETE favorite/{favoriteId}"
      operationId: "deleteFavorite"
      parameters:
        - name: "favoriteId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET favorite/{favoriteId}"
      operationId: "getFavoriteById"
      parameters:
        - name: "favoriteId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
  /favorite/:
    post:
      summary: "POST favorite/"
      operationId: "createFavorite"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
    get:
      summary: "GET favorite/"
      operationId: "getAllFavorites"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEntity"
...
components:
  schemas:
    LoginRequestDTO:
      type: "object"
      properties:
        email:
          type: "string"
        password:
          type: "string"
    HttpHeaders:
      type: "object"
      properties: { }
    ResponseEntity:
      type: "object"
      properties:
        headers:
          $ref: "#/components/schemas/HttpHeaders"
      required:
        - "headers"
    RegisterRequestDTO:
      type: "object"
      properties:
        name:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        avatar:
          type: "array"
          items:
            type: "string"
            format: "byte"
        createdAt:
          type: "string"
        status:
          type: "boolean"
        role:
          type: "string"
    EnsembleRequestDTO:
      type: "object"
      properties:
        name:
          type: "string"
        description:
          type: "string"
        userId:
          type: "integer"
          format: "int32"
    User:
      type: "object"
      properties:
        userId:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        password:
          type: "string"
        email:
          type: "string"
        role:
          type: "string"
        avatar:
          type: "array"
          items:
            type: "string"
            format: "byte"
        createdAt:
          type: "string"
        updatedAt:
          type: "string"
        status:
          type: "boolean"
    Ensemble:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        description:
          type: "string"
        cover:
          type: "array"
          items:
            type: "string"
            format: "byte"
        user:
          $ref: "#/components/schemas/User"
        status:
          type: "boolean"
...
    Flashcard:
      type: "object"
      properties:
        flashcardId:
          type: "integer"
          format: "int32"
        title:
          type: "string"
        content:
          type: "string"
        createdAt:
          type: "string"
        updatedAt:
          type: "string"
        ensemble:
          $ref: "#/components/schemas/Ensemble"
        status:
          type: "boolean"
...
    BoxRequestDTO:
      type: "object"
      properties:
        boxId:
          type: "integer"
          format: "int64"
        price:
          type: "integer"
          format: "int32"
        title:
          type: "string"
        cover:
          type: "array"
          items:
            type: "string"
            format: "byte"
        description:
          type: "string"
        status:
          type: "boolean"
        userId:
          type: "integer"
          format: "int32"
    FavoriteRequestDTO:
      type: "object"
      properties:
        farvoriteId:
          type: "integer"
          format: "int64"
        favoritationDateTimestamp:
          type: "string"
        status:
          type: "boolean"
        userId:
          type: "integer"
          format: "int32"
...